pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.1",
    "networkx>=3.4.2",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pyinstaller>=6.12.0",
    "scapy>=2.6.1",
    "streamlit>=1.44.0",
    "customtkinter>=5.2.2",
    "ttkthemes>=3.2.2",
    "psutil>=7.0.0",
    "setuptools>=78.1.0",
    "kamene>=0.32",
    "scapy-python3>=0.26",
    "pillow>=11.1.0",
]
